jetpack: re organize QRPost plugin
